// cli/src/core/api/types.ts
export interface PaginatedResponse<T> {
  totalPages: number;
  totalElements: number;
  data: T[];
}

export interface PaginationParams {
  limit?: number;
  offset?: number;
}

export interface SearchParams extends PaginationParams {
  filter?: Record<string, unknown>;
  sort?: { sortBy: string[]; order: 'ASC' | 'DESC' };
}
// ...
export interface FixedAsset {
  resourceId: string;
  name: string;
  reference: string;
  typeName: string;
  category: string;
  purchaseDate: number;
  status: string;
  purchaseAmount: number;
  netBookValueAmount: number;
  registrationType: string;
  disposalValueDate: number | null;
  assetDisposalGainLossAmount: number;
  netBookAtDisposalAmount: number;
  bookValueAmount: number;
  internalNotes: string | null;
  customFields?: Record<string, unknown>[];
}